ref(event-schema): Merge trace data with span data
