+++ relay-event-schema/src/protocol/span/convert.rs
@@ -207,12 +207,12 @@ mod tests {
                 user_agent_original: ~,
                 url_full: ~,
                 client_address: ~,
+                route: ~,
+                previous_route: ~,
                 other: {
                     "custom_attribute": I64(
                         42,
                     ),
-                    "previousRoute": ~,
-                    "route": ~,
                 },
             },
             sentry_tags: ~,
